requirements.txt
psycopg2-binary
python-dotenv
pyrogram
aiogram
apscheduler
fpdf
pillow
dropbox
